software/desktop/thermal_display/requirements.txt
adafruit-ampy==1.1.0
args==0.1.0
bitarray==2.8.2
bitstring==4.1.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clint==0.5.1
colour==0.1.5
contourpy==1.3.1
cryptography==41.0.4
cycler==0.12.1
dacite==1.8.1
docopt==0.6.2
ecdsa==0.18.0
esptool==4.6.2
fonttools==4.56.0
idna==3.7
kiwisolver==1.4.8
matplotlib==3.10.0
mpremote==1.24.0
mutagen==1.47.0
numpy==2.1.1
packaging==24.2
paho-mqtt==2.1.0
pathvalidate==3.2.0
pillow==11.0.0
pycparser==2.21
pygame==2.6.1
pyparsing==3.2.1
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pyudev==0.24.1
PyYAML==6.0.1
reedsolo==1.7.0
requests==2.32.3
rshell==0.0.31
scdl==2.7.9
scipy==1.15.2
setuptools==68.2.2
six==1.16.0
soundcloud-v2==1.3.7
termcolor==2.4.0
urllib3==2.2.1
